requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
# backports.zoneinfo
bcrypt==4.1.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
cssselect2==0.7.0
dj-database-url==0.5.0
Django==4.2.17
django-bootstrap3==21.2
django-mail-templated==2.6.5
django-utils-six==2.0
django-widget-tweaks==1.5.0
gunicorn==22.0.0
html5lib==1.1
idna==3.7
invoke==2.2.0
lxml
# mysql-client==0.0.1
# mysql-connector-python==9.0.0
oscrypto==1.3.0
packaging==24.1
paramiko==3.4.0
pillow==10.4.0
pycparser==2.22
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyMySQL==1.1.1
PyNaCl==1.5.0
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
xhtml2pdf==0.2.16
